Upload email sent to AD
Formulated the hypothesis as a data science problem and laid out the major steps needed to test the hypothesis. Communicated thoughts and findings in email to AD, focusing on the data that you would need from the client and the analytical models you would use to test such a hypothesis.